fix: web vitals null value | move uptime and reachability to common
