+++ src/scenes/BrowserDashboard/BrowserDashboard.tsx
@@ -19,8 +19,8 @@ import { DataSent } from 'scenes/Common/DataSent';
 import { DistinctTargets } from 'scenes/Common/DistinctTargets';
 import { DurationByProbe } from 'scenes/Common/DurationByProbe';
 import { ErrorLogs } from 'scenes/Common/ErrorLogsPanel';
-import { ReachabilityStat } from 'scenes/HTTP/stats/ReachabilityStatViz';
-import { UptimeStat } from 'scenes/HTTP/stats/UptimeStatViz';
+import { ReachabilityStat } from 'scenes/Common/ReachabilityStatViz';
+import { UptimeStat } from 'scenes/Common/UptimeStatViz';
 
 export const BrowserDashboard = ({ check }: { check: Check }) => {
   const styles = useStyles2(getStyles);
